refactor: migrate sign-up flow to client side
